debug_tool/requirements.txt
pyserial>=3.5
pyusb>=1.2.1
pywebview>=4.4
PySide6>=6.7
requests>=2.25.0
Pillow>=8.0.0